% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_content.R
\name{render_exercises}
\alias{render_exercises}
\title{Deploy the exercises in HTML format}
\usage{
render_exercises(exercise, solved = FALSE)
}
\arguments{
\item{exercise}{character object indicating which exercise to visualise:
'Exercise1', 'Exercise2', or 'Exercise3'}

\item{solved}{logical argument. Should the solution be displayed? The default is FALSE}
}
\value{
this function directs the user to the selected exercise.
}
\description{
Deploy the exercises in HTML format
}
