% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_content.R
\name{render_lectures}
\alias{render_lectures}
\title{Deploy the lecture slides in HTML format}
\usage{
render_lectures(lecture)
}
\arguments{
\item{lecture}{character object indicating which lecture to visualise:
\itemize{
  \item{'Lecture1': }{Introduction to R - Part 1}
  \item{'Lecture2': }{Introduction to R - Part 2}
  \item{'Lecture3': }{Raster files and spatial data}
  \item{'Lecture4': }{Data processing}
  \item{'Lecture5': }{Spatial and temporal statistics}
  \item{'Lecture6': }{Automating downloading}
}}
}
\value{
this function directs the user to the selected lecture.
}
\description{
Deploy the lecture slides in HTML format
}
